Move http related utils in an own commons package
