+++ jutrack-core/src/main/java/com/github/ibiber/commons/http/UrlParameterBuilder.java
@@ -1,4 +1,4 @@
-package com.github.ibiber.jutrack.util;
+package com.github.ibiber.commons.http;
 
 import java.util.ArrayList;
 import java.util.List;
